🎨  refactor to take a single master config, allow passing in custom shopStore strategies
