+++ middleware/index.js
@@ -1,0 +1,12 @@
+const createWithWebhook = require('./webhooks');
+const createWithShop = require('./withShop');
+
+module.exports = function createMiddleware(shopifyConfig) {
+  const withWebhook = createWithWebhook(shopifyConfig)
+  const withShop = createWithShop();
+
+  return {
+    withShop,
+    withWebhook,
+  };
+};
