--- shopStore/index.js
@@ -1,3 +1,0 @@
-const Store = require('./ShopStore');
-
-module.exports = new Store(process.env.SHOPIFY_APP_STORAGE_ENGINE);
