+++ strategies/SqliteStrategy.js
