+++ strategies/index.js
@@ -1,0 +1,9 @@
+const RedisStrategy = require('./RedisStrategy');
+const MemoryStrategy = require('./MemoryStrategy');
+const SqliteStrategy = require('./SqliteStrategy');
+
+module.exports = {
+  RedisStrategy,
+  MemoryStrategy,
+  SqliteStrategy,
+}
